+++ src/clients/api/mutations/repayNonBnbVToken.ts
@@ -4,19 +4,19 @@ import type { TransactionReceipt } from 'web3-core/types';
 import { VTokenId } from 'types';
 import { VTokenContract } from 'clients/contracts/types';
 
-export interface IRepayNonBnbInput {
+export interface IRepayNonBnbVTokenInput {
   vTokenContract: VTokenContract<Exclude<VTokenId, 'bnb'>>;
   fromAccountAddress: string;
   amountWei: BigNumber;
 }
 
-export type RepayNonBnbOutput = TransactionReceipt;
+export type RepayNonBnbVTokenOutput = TransactionReceipt;
 
-const repayNonBnb = async ({
+const repayNonBnbVToken = async ({
   vTokenContract,
   fromAccountAddress,
   amountWei,
-}: IRepayNonBnbInput): Promise<RepayNonBnbOutput> =>
+}: IRepayNonBnbVTokenInput): Promise<RepayNonBnbVTokenOutput> =>
   vTokenContract.methods.repayBorrow(amountWei.toFixed()).send({ from: fromAccountAddress });
 
-export default repayNonBnb;
+export default repayNonBnbVToken;
