+++ src/clients/api/mutations/useRepayVToken.ts
@@ -2,28 +2,28 @@ import { MutationObserverOptions } from 'react-query';
 
 import { VTokenId } from 'types';
 import {
-  useRepayNonBnb,
+  useRepayNonBnbVToken,
   useRepayBnb,
   RepayBnbOutput,
-  RepayNonBnbOutput,
-  IRepayNonBnbInput,
+  RepayNonBnbVTokenOutput,
+  IRepayNonBnbVTokenInput,
   IRepayBnbInput,
 } from 'clients/api';
 
 type Options = MutationObserverOptions<
-  RepayBnbOutput | RepayNonBnbOutput,
+  RepayBnbOutput | RepayNonBnbVTokenOutput,
   Error,
-  Omit<IRepayNonBnbInput, 'vTokenContract'> | Omit<IRepayBnbInput, 'web3'>
+  Omit<IRepayNonBnbVTokenInput, 'vTokenContract'> | Omit<IRepayBnbInput, 'web3'>
 >;
 
 const useRepayVToken = ({ vTokenId }: { vTokenId: VTokenId }, options?: Options) => {
-  const useRepayNonBnbResult = useRepayNonBnb(
+  const useRepayNonBnbVTokenResult = useRepayNonBnbVToken(
     { vTokenId: vTokenId as Exclude<VTokenId, 'bnb'> },
     options,
   );
   const useRepayBnbResult = useRepayBnb(options);
 
-  return vTokenId === 'bnb' ? useRepayBnbResult : useRepayNonBnbResult;
+  return vTokenId === 'bnb' ? useRepayBnbResult : useRepayNonBnbVTokenResult;
 };
 
 export default useRepayVToken;
